Course routing system Fetching json from folder again
