+++ package.json
@@ -20,6 +20,7 @@
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-icons": "^4.3.1",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
     "typescript": "^4.6.4",
     "web-vitals": "^2.1.4"
