+++ src/App.tsx
@@ -1,7 +1,17 @@
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
+import CurriculumPage from './pages/Curriculum';
 
 function App(): React.ReactElement {
-  return <Home />;
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="*" element={<Home />} />
+        <Route path="/:university/:course" element={<CurriculumPage />} />
+      </Routes>
+    </BrowserRouter>
+  );
 }
 
 export default App;
