+++ src/models/Curriculum.ts
@@ -6,12 +6,13 @@ interface IpreRequisites {
 export type DisciplineType = {
   name: string;
   code: string;
-  prerequisites: IpreRequisites[];
-  semester: number;
   hours: number;
   credits: number;
-  isObligatory: true;
-  isActive: boolean;
+  semester: number;
+  isActive?: boolean;
+  isObligatory: boolean;
+  equivalences: IpreRequisites[];
+  prerequisites: IpreRequisites[];
 };
 
 export type CurriculumType = {
@@ -19,9 +20,9 @@ export type CurriculumType = {
   course: string;
   semesters: number;
   totalHours: number;
-  totalHoursObligatory: number;
   totalHoursElective: number;
-  disciplines: [DisciplineType];
+  totalHoursObligatory: number;
+  disciplines: DisciplineType[];
 };
 
 export interface Itime {
@@ -30,11 +31,11 @@ export interface Itime {
   eletiva: number;
 }
 
-export type hoursSetterFuncTypes = {
+export type HoursSetterFuncTypes = {
   setAcademicObligatoryDone: React.Dispatch<React.SetStateAction<number>>;
   setAcademicElectiveDone: React.Dispatch<React.SetStateAction<number>>;
   setAcademicTotalDone: React.Dispatch<React.SetStateAction<number>>;
-}
+};
 
 export type OnClickTypes = {
   isActive: boolean;
