+++ src/utils/requestDisciplines.ts
@@ -1,6 +1,10 @@
 import { CurriculumType } from '../models/Curriculum';
 
-const disciplineData = fetch('http://localhost:3000/api').then<CurriculumType>(
-  (response) => response.json()
-);
-export default disciplineData;
+function requestDisciplines(university: string, course: string) {
+  const disciplineData: Promise<CurriculumType> = fetch(
+    `${process.env.PUBLIC_URL}/university/${university}/${course}.json`
+  ).then((response) => response.json());
+  return disciplineData;
+}
+
+export default requestDisciplines;
